LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/abspos-in-multicol-001-expected.txt

abcd efg def ghi

FAIL Anchoring with text wrapping changes. assert_equals: expected 250 but got 150
